+++ src/handlebars_vm.c
@@ -330,6 +330,7 @@ static inline struct handlebars_value * execute_template(
     vm2->flags = vm->flags;
     vm2->helpers = vm->helpers;
     vm2->partials = vm->partials;
+    vm2->data = vm->data; // need to pass data along to partials
 
     // Copy stacks
     memcpy(&vm2->contextStack, &vm->contextStack, offsetof(struct handlebars_vm_stack, v) + (sizeof(struct handlebars_value *) * LEN(vm->contextStack)));
@@ -351,6 +352,7 @@ static inline struct handlebars_value * execute_template(
         vm->cache->release(vm->cache, tmpl, module);
     }
 
+    vm2->data = NULL;
     handlebars_vm_dtor(vm2);
 
 // done:
@@ -614,7 +616,6 @@ ACCEPT_FUNCTION(invoke_partial)
     struct handlebars_value * tmp = NULL;
     struct handlebars_string * name = NULL;
     struct handlebars_value * partial = NULL;
-    jmp_buf buf;
     int argc = 1;
     struct handlebars_value * argv[1];
 
@@ -648,13 +649,15 @@ ACCEPT_FUNCTION(invoke_partial)
         if( vm->flags & handlebars_compiler_flag_compat ) {
             return;
         } else {
-            handlebars_throw(CONTEXT, HANDLEBARS_ERROR, "The partial %s could not be found", name ? name->val : "(NULL)");
+            if (!name) {
+                name = handlebars_string_ctor(CONTEXT, HBS_STRL("(NULL)"));
+            }
+            handlebars_throw(CONTEXT, HANDLEBARS_ERROR, "The partial %.*s could not be found", (int) name->len, name->val);
         }
     }
 
     // Construct new context
     struct handlebars_context * context = handlebars_context_ctor_ex(vm);
-    context->e->jmp = &buf;
 
     // If partial is a function?
     if( handlebars_value_is_callable(partial) ) {